feat: new jubeat tables and folders
