+++ client/src/index.tsx
@@ -2,10 +2,10 @@ import "@fortawesome/fontawesome-free/css/all.min.css";
 import React from "react";
 import ReactDOM from "react-dom";
 import App from "./app/App";
+import "./_style/base.scss";
 import "./_assets/flaticon/flaticon.css";
 import "./_assets/flaticon2/flaticon.css";
 import "./_assets/keenthemes-icons/font/ki.css";
-import "./_style/base.scss";
 
 const { PUBLIC_URL } = process.env;
